refactor: adapta caminho para ícones svg
